Backend/requirements.txt
# Core LangChain packages
langchain == 0.3.27
langchain-community == 0.3.31
langchain-text-splitters ==  0.3.11
langchain-huggingface ==  0.3.1
langchain-mongodb == 0.7.1
langchain-groq == 0.3.8

# Embeddings
sentence-transformers == 5.1.1

python-multipart

# Authentication
bcrypt==4.1.3
PyJWT==2.9.0
requests==2.32.5
google-auth

# NLP utilities
nltk == 3.9.2

# Document parsing
unstructured == 0.18.15
unstructured[pdf]

# Environment variable management
python-dotenv==1.0.1

pydantic[email]==2.8.2
email-validator==2.0.0

fastapi
uvicorn[standard]
pymongo

